+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/CityTablePanel.java
@@ -90,7 +90,7 @@ final class CityTablePanel extends ChartTablePanel {
     }
 
     @Override
-    public Void perform(ProgressReporter<String> progressReporter) throws Exception {
+    public Void execute(ProgressReporter<String> progressReporter) throws Exception {
       tableModel.populateLocationForSelected(progressReporter, cancelled);
       return null;
     }