+++ swing/common-ui-tools/src/main/java/is/codion/swing/common/ui/tools/loadtest/LoadTestPanel.java
@@ -501,7 +501,7 @@ public final class LoadTestPanel<T> extends JPanel {
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       scenario.clearExceptions();
       exceptionsTextArea.replaceRange("", 0, exceptionsTextArea.getDocument().getLength());
     }
@@ -518,7 +518,7 @@ public final class LoadTestPanel<T> extends JPanel {
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       exceptionsTextArea.replaceRange("", 0, exceptionsTextArea.getDocument().getLength());
       for (Throwable exception : scenario.exceptions()) {
         exceptionsTextArea.append(exception.getMessage());