+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DefaultOkCancelDialogBuilder.java
@@ -153,7 +153,7 @@ final class DefaultOkCancelDialogBuilder extends DefaultActionDialogBuilder<OkCa
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       command.run();
       Utilities.disposeParentWindow(component);
     }
@@ -168,7 +168,7 @@ final class DefaultOkCancelDialogBuilder extends DefaultActionDialogBuilder<OkCa
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       Utilities.disposeParentWindow(component);
     }
   }
@@ -182,7 +182,7 @@ final class DefaultOkCancelDialogBuilder extends DefaultActionDialogBuilder<OkCa
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       Utilities.disposeParentWindow(component);
     }
   }