+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/Dialogs.java
@@ -67,7 +67,7 @@ public final class Dialogs {
     requireNonNull(task);
 
     return new DefaultProgressWorkerDialogBuilder<>(progressReporter -> {
-      task.perform();
+      task.execute();
       return null;
     }).indeterminate(true);
   }
@@ -80,7 +80,7 @@ public final class Dialogs {
   public static <T> ProgressWorkerDialogBuilder<T, ?> progressWorkerDialog(ProgressWorker.Task<T> task) {
     requireNonNull(task);
 
-    return new DefaultProgressWorkerDialogBuilder<>(progressReporter -> task.perform()).indeterminate(true);
+    return new DefaultProgressWorkerDialogBuilder<>(progressReporter -> task.execute()).indeterminate(true);
   }
 
   /**