+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/TabbedPanelLayout.java
@@ -561,7 +561,7 @@ public final class TabbedPanelLayout implements PanelLayout {
       }
 
       @Override
-      public void perform() {
+      public void execute() {
         if (panelState.equalTo(HIDDEN)) {
           panelState.set(EMBEDDED);
         }