+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityControls.java
@@ -203,7 +203,7 @@ final class EntityControls {
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       EntityEditPanel editPanel = initializeEditPanel();
       editPanel.editModel().setDefaults();
       State cancelled = State.state();
@@ -313,7 +313,7 @@ final class EntityControls {
     }
 
     @Override
-    public void perform() throws Exception {
+    public void execute() throws Exception {
       if (component instanceof EntityComboBox) {
         if (((EntityComboBox) component).isPopupVisible()) {
           ((EntityComboBox) component).hidePopup();