+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -717,7 +717,7 @@ public final class EntitySearchField extends HintTextField {
       }
 
       @Override
-      public void perform() {
+      public void execute() {
         searchModel.entities().set(list.getSelectedValuesList());
         Utilities.disposeParentWindow(list);
       }